+++ R/tools.R
@@ -625,11 +625,11 @@ simplify_text <- function(text)  {
 #============================================================================
 #' Rescales a numeric variable into values between 0 and 1
 #'
-#' @param var text string
-#' @return text string
+#' @param x numeric vector (to be rescaled)
+#' @return vector with values between 0 and 1
 #' @import dplyr
 #' @examples
-#' simplify_text(" Hello  World !, ")
+#' rescale01(0:10)
 #' @export
 
 rescale01 <- function(x)  {
@@ -745,3 +745,35 @@ clean_var <- function(data, var, na = NA, min_val = NA, max_val = NA, max_cat = 
   # return data
   data
 } # clean_var
+
+#============================================================================
+#  count_pct
+#============================================================================
+#' Adds percentage to dplyr::count()
+#'
+#' Adds variables tot (total) and
+#' pct (percentage) to dplyr::count()
+#'
+#' @param data A dataset
+#' @param ... Other parameters passed to count()
+#' @return Dataset
+#' @import dplyr
+#' @examples
+#' count_pct(iris, Species)
+#' @export
+
+count_pct <- function(data, ...)  {
+
+  d <- data %>%
+    count(...)
+
+  names(d) <- c("value", "n")
+
+  d <- d %>%
+    mutate(tot = sum(n),
+           pct = n / sum(n) * 100.00)
+
+  d
+} # count_pct
+
+